Converting to npm package
