--- .codeclimate.yml
@@ -1,17 +1,0 @@
-engines:
-  eslint:
-    enabled: true
-  csslint:
-    enabled: true
-
-ratings:
-  paths:
-    - src/**
-
-exclude_paths:
-  - app/*
-  - dist/*
-  - tests/*
-  - build/*
-  - karma.conf.js
-  - gulpfile.js
