--- .editorconfig
@@ -1,18 +1,0 @@
-# http://editorconfig.org
-
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 4
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.md]
-trim_trailing_whitespace = false
-
-[{package.json,bower.json,.travis.yml}]
-indent_style = space
-indent_size = 2
