--- .jshintrc
@@ -1,52 +1,0 @@
-{
-  /*
-   * ENVIRONMENTS
-   * =================
-   */
-
-  // Define globals exposed by modern browsers.
-  "browser": true,
-
-  /*
-   * ENFORCING OPTIONS
-   * =================
-   */
-
-  // Prohibit use of == and != in favor of === and !==.
-  "eqeqeq": true,
-
-  // Prohibit use of a variable before it is defined.
-  "latedef": true,
-
-  // Require capitalized names for constructor functions.
-  "newcap": true,
-
-  // Enforce use of single quotation marks for strings.
-  "quotmark": "single",
-
-  // Prohibit use of explicitly undeclared variables.
-  "undef": true,
-
-  // Warn when variables are defined but never used.
-  "unused": true,
-
-  // Require {} for every new block or scope
-  "curly": true,
-
-  /*
-   * RELAXING OPTIONS
-   * =================
-   */
-
-  // Suppress warnings about == null comparisons.
-  "eqnull": true,
-
-  /*
-   * CUSTOM GLOBALS
-   * =================
-   */
-
-  "globals": {
-    "angular": true
-  }
-}
