+++ .npmrc
@@ -1,0 +1,2 @@
+@rally:registry=https://artifacts.werally.in/artifactory/api/npm/npm-local
+always-auth=true
