--- app/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <meta name="description" content="Sample App">
-        <meta name="author" content="Rahul Doshi">
-
-        <title>My AngularJS App</title>
-
-        <!-- Bootstrap core CSS -->
-        <link href="../lib/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
-
-        <link href="../lib/bca-flag-sprite/css/flags.css" rel="stylesheet">
-
-        <!-- Custom styles for this template -->
-        <link href="styles/custom.css" rel="stylesheet">
-    </head>
-
-    <body data-ng-app="myApp">
-        <div data-ng-include="'src/shared/components/navbar/_navbar.html'"></div>
-        <div data-ng-view></div>
-        <div data-ng-include="'src/shared/components/footer/_footer.html'"></div>
-
-        <!-- Bootstrap core JavaScript
-        ================================================== -->
-        <!-- Placed at the end of the document so the pages load faster -->
-        <script src="../lib/jquery/dist/jquery.min.js"></script>
-        <script src="../lib/bootstrap/dist/js/bootstrap.min.js"></script>
-
-        <script src="../lib/angular/angular.js"></script>
-        <script src="../lib/angular-route/angular-route.min.js"></script>
-        <script src="../lib/angular-cookies/angular-cookies.min.js"></script>
-        <script src="../lib/angular-sanitize/angular-sanitize.min.js"></script>
-
-        <script src="../dist/angular-localization.js"></script>
-
-        <script src="src/core/declarations.js"></script>
-        <script src="src/core/routes.js"></script>
-
-        <script src="src/shared/components/footer/footer.ctrl.js"></script>
-
-        <script src="src/features/demo/demo.route.js"></script>
-    </body>
-</html>
