--- app/languages/en-US/common.lang.json
@@ -1,4 +1,0 @@
-{
-    "yes": "Yes",
-    "no": "No"
-}
