--- app/languages/en-US/demo.lang.json
@@ -1,7 +1,0 @@
-{
-    "emailAddress": "Email Address",
-    "emailPlaceholder": "Enter Email",
-    "formDataMessage": "Your form data will remain intact when changing languages. Go ahead and try it out!",
-    "password": "Password",
-    "submit": "Submit"
-}
