--- app/languages/en-US/footer.lang.json
@@ -1,3 +1,0 @@
-{
-    "copyrightString": "© {currYear} {fullName} All rights reserved."
-}
