--- app/languages/en-US/navbar.lang.json
@@ -1,4 +1,0 @@
-{
-    "demo": "Demo",
-    "home": "Home"
-}
