--- app/languages/fr-FR/common.lang.json
@@ -1,4 +1,0 @@
-{
-    "yes": "Oui",
-    "no": "Aucun"
-}
