--- app/languages/fr-FR/demo.lang.json
@@ -1,7 +1,0 @@
-{
-    "emailAddress": "Adresse e-mail",
-    "emailPlaceholder": "Entrer email",
-    "formDataMessage": "Vos données de formulaire resteront intactes lors du changement de langues. Allez-y et essayez-le!",
-    "password": "Mot de passe",
-    "submit": "Soumettre"
-}
