--- app/languages/fr-FR/footer.lang.json
@@ -1,3 +1,0 @@
-{
-    "copyrightString": "© {currYear} {fullName} Tous droits réservés."
-}
