--- app/languages/fr-FR/navbar.lang.json
@@ -1,4 +1,0 @@
-{
-    "demo": "Démo",
-    "home": "Maison"
-}
