--- app/src/core/views/_home.html
@@ -1,11 +1,0 @@
-<!-- Begin page content -->
-<div class="container">
-    <div class="page-header">
-        <h1>angular-localization</h1>
-    </div>
-    <p class="lead">
-        A complete AngularJS localization service with core service and accompanying filters, directives etc.
-        <br>
-        Site localization example using the <code>ngLocalize</code> module.
-    </p>
-</div>
