--- app/src/features/demo/demo.route.js
@@ -1,12 +1,0 @@
-angular.module('myApp.features.demo')
-    .config(['$routeProvider',
-        function ($routeProvider) {
-            'use strict';
-
-            $routeProvider.
-                when('/demo', {
-                    templateUrl: 'src/features/demo/demo.part.html',
-                    // controller: 'HomeControl'
-                });
-        }
-    ]);
