--- app/src/shared/components/navbar/_navbar.html
@@ -1,20 +1,0 @@
-<!-- Fixed navbar -->
-<div class="navbar navbar-default navbar-fixed-top" role="navigation">
-    <div class="container">
-        <div class="navbar-header">
-            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
-                <span class="sr-only">Toggle navigation</span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-            </button>
-            <a class="navbar-brand" href="#">Sample App</a>
-        </div>
-        <div class="collapse navbar-collapse">
-            <ul class="nav navbar-nav">
-                <li><a href="#home">{{ 'navbar.home' | i18n }}</a></li>
-                <li><a href="#demo">{{ 'navbar.demo' | i18n }}</a></li>
-            </ul>
-        </div><!--/.nav-collapse -->
-    </div>
-</div>
