--- bower.json
@@ -1,45 +1,0 @@
-{
-  "name": "angular-localization",
-  "version": "1.5.1",
-  "homepage": "https://github.com/doshprompt/angular-localization",
-  "description": "angularjs localization done right",
-  "main": [
-    "dist/angular-localization.js",
-    "dist/angular-localization.css"
-  ],
-  "keywords": [
-    "angularjs",
-    "localization",
-    "i18n",
-    "l10n",
-    "internationalization",
-    "angular",
-    "ng"
-  ],
-  "authors": [
-    "Rahul Doshi"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests",
-    "src",
-    "app",
-    "build",
-    "gulpfile.js"
-  ],
-  "dependencies": {
-    "angular": ">=1.2 < 2",
-    "angular-cookies": ">=1.2 < 2",
-    "angular-sanitize": ">=1.2 < 2"
-  },
-  "devDependencies": {
-    "angular-mocks": ">=1.2 < 2",
-    "bootstrap": "~3.3.4",
-    "angular-route": ">=1.2 < 2",
-    "bca-flag-sprite": "~1.0.0"
-  }
-}
