+++ index.js
@@ -1,0 +1,2 @@
+require('./angular-localization');
+module.exports = 'ngLocalize';
