--- module.prefix
@@ -1,2 +1,0 @@
-;(function (angular, window, document, undefined) {
-    'use strict';
