--- module.suffix
@@ -1,1 +1,0 @@
-}(this.angular, this, this.document));
