+++ package.json
@@ -1,18 +1,11 @@
 {
-  "name": "angular-localization",
+  "name": "starship-angular-localization-npm",
   "version": "1.5.1",
   "description": "angularjs localization done right.",
-  "main": [
-    "dist/angular-localization.js",
-    "dist/angular-localization.css"
-  ],
-  "scripts": {
-    "postinstall": "bower install",
-    "test": "gulp test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/doshprompt/angular-localization.git"
+  "main": "index.js",
+  "repository": "https://github.com/AudaxHealthInc/starship-angular-localization-npm.git",
+  "publishConfig": {
+    "registry": "https://artifacts.werally.in/artifactory/api/npm/npm-local"
   },
   "keywords": [
     "angularjs"
@@ -23,33 +16,6 @@
     "url": "https://github.com/doshprompt/angular-localization/issues"
   },
   "homepage": "http://doshprompt.github.io/angular-localization",
-  "devDependencies": {
-    "bower": "^1.4.1",
-    "codeclimate-test-reporter": "^0.1.0",
-    "del": "^0.1.3",
-    "gulp": "^3.9.0",
-    "gulp-add-src": "^0.2.0",
-    "gulp-angular-filesort": "^1.1.1",
-    "gulp-concat": "^2.4.1",
-    "gulp-header": "^1.1.1",
-    "gulp-jshint": "^1.11.2",
-    "gulp-less": "^3.0.3",
-    "gulp-load-plugins": "^0.10.0",
-    "gulp-ng-annotate": "^0.5.2",
-    "gulp-preprocess": "^1.1.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^1.5.2",
-    "gulp-uglify": "^1.2.0",
-    "gulp-webserver": "^0.9.1",
-    "jshint-stylish": "^2.0.1",
-    "karma": "^0.12.24",
-    "karma-coverage": "^0.4.2",
-    "karma-jasmine": "^0.1.5",
-    "karma-junit-reporter": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "less-plugin-clean-css": "^1.5.0",
-    "moment": "^2.8.3",
-    "phantomjs-prebuilt": "^2.1.4",
-    "run-sequence": "^1.1.1"
-  }
+  "dependencies": {},
+  "devDependencies": {}
 }
