--- src/localization.config.js
@@ -1,12 +1,0 @@
-angular.module('ngLocalize.Config', [])
-    .value('localeConf', {
-        basePath: 'languages',
-        defaultLocale: 'en-US',
-        sharedDictionary: 'common',
-        fileExtension: '.lang.json',
-        persistSelection: true,
-        cookieName: 'COOKIE_LOCALE_LANG',
-        observableAttrs: new RegExp('^data-(?!ng-|i18n)'),
-        delimiter: '::',
-        validTokens: new RegExp('^[\\w\\.-]+\\.[\\w\\s\\.-]+\\w(:.*)?$')
-    });
