--- src/localization.events.js
@@ -1,5 +1,0 @@
-angular.module('ngLocalize.Events', [])
-    .constant('localeEvents', {
-        resourceUpdates: 'ngLocalizeResourcesUpdated',
-        localeChanges: 'ngLocalizeLocaleChanged'
-    });
