--- src/localization.helpers.less
@@ -1,9 +1,0 @@
-// Text Overflow Truncate
-// -------------------------
-
-.overflow-ellipsis {
-    width: 100%;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
