--- src/localization.langs.js
@@ -1,7 +1,0 @@
-angular.module('ngLocalize.InstalledLanguages', [])
-    .value('localeSupported', [
-        'en-US'
-    ])
-    .value('localeFallbacks', {
-        'en': 'en-US'
-    });
