--- src/localization.module.js
@@ -1,1 +1,0 @@
-angular.module('ngLocalize', ['ngSanitize', 'ngLocalize.Config', 'ngLocalize.Events', 'ngLocalize.InstalledLanguages']);
