--- src/localization.version.js
@@ -1,2 +1,0 @@
-angular.module('ngLocalize.Version', [])
-    .constant('localeVer', '/* @echo VERSION */');
