--- tests/e2e/scenarios.js
@@ -1,9 +1,0 @@
-describe('my app', function() {
-    'use strict';
-
-    browser.get('index.html');
-
-    it('should automatically redirect to /homw when location hash/fragment is empty', function() {
-        expect(browser.getLocationAbsUrl()).toMatch("/home");
-    });
-});
